requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pytorch-ignite==0.4.9
requests==2.28.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
typing_extensions==4.2.0
urllib3==1.26.9
